feat: introduce a wip implementation of the inertiaframe (#1)
interiaframe is an opinionated framework on top of the Inertia for a protocol-free communication. The primary idea behind inertiaframe is that HTTP and Inertia layers should be abstracted as much as possible, yet allowing for an easy hijacking for a lower-level control.

Communication in interiaframe is primarily handled via message passing, the incoming request is parsed into an expected request struct and optionally validated. Response is a specific response object created via a NewResponse with the component name to render and the props, generally, passed as a struct.